Fix errors tracebacks during system dataset migration
Export of pool and related system dataset migration triggered
two tracebacks. One regression in smb_configure due to regression
caused by groupmap_list output format changes and the other
in reporting setup due to failure to check whether symlink
already exists.

+++ src/middlewared/middlewared/plugins/reporting.py
@@ -890,7 +890,7 @@ class ReportingService(ConfigService):
             os.makedirs(os.path.join(pwd, 'localhost'))
 
         # Create "${hostname}" -> "localhost" symlink if necessary
-        if hostname != 'localhost':
+        if hostname != 'localhost' and not os.path.exists(os.path.join(pwd, hostname)):
             os.symlink(os.path.join(pwd, 'localhost'), os.path.join(pwd, hostname))
 
         # Let's return a positive value to indicate that necessary collectd operations were performed successfully
